+++ src/strategies/DrivingStrategy.java
@@ -4,10 +4,10 @@ public interface DrivingStrategy {
     void accelerate();
     void handleRoadConditions();
     void navigate();
-    double calculateFuelConsumption(double distance);
+    float calculateFuelConsumption(float distance);
     String getModeName();
 
-    default double formatFuelConsumption(double fuel) {
-        return Math.round(fuel * 100.0) / 100.0;
+    default float formatFuelConsumption(float fuel) {
+        return Math.round(fuel * 100.0f) / 100.0f;
     }
 }
